feat: restructure pkp-cosmos
