+++ packages/lit-third-party-libs/src/index.ts
@@ -1,2 +1,2 @@
 export { IPFSBundledSDK } from './lib/ipfs-bundled-sdk';
-export { CosmosBundledSDK } from './lib/cosmos-bundled-sdk';
+// export { CosmosBundledSDK } from './lib/cosmos-bundled-sdk';
