+++ packages/pkp-cosmos/src/index.ts
@@ -1,6 +1,6 @@
 // ----- autogen:polyfills:start  -----
 //
 // ----- autogen:polyfills:end  -----
-import { PKPCosmosWallet } from './lib/pkp-cosmos-wallet';
+import { PKPCosmosWallet } from './lib/pkp-cosmos';
 
 export { PKPCosmosWallet };
